src/mlx-openai-optiq-server/requirements.txt
mlx-lm>=0.30.7
mlx-optiq
fastapi
uvicorn
pydantic
